Move enum to constant folder
