+++ practice/django_practice/employees/constant/enum.py
@@ -1,0 +1,12 @@
+from enum import Enum
+
+
+class ActiveStatus(Enum):
+    ACTIVE = "Active"
+    DISABLED = "Disabled"
+
+
+class Gender(Enum):
+    MALE = "Male"
+    FEMALE = "Female"
+    OTHER = "Other"
